feat: Disable SSR when developing
BREAKING CHANGE: SSR is no longer used in development mode. This has the major benefit of being faster (no rendering of React), obviously. In the client side ReactDOM.hydrate is only used in production, while ReactDOM.render is used when developing.

Another benefit is that previously, the server was run via @babel/node and restarted with nodemon on changes to server code. This slowed rebuilding and also wasn’t perferct. It didn’t detect changes in the client code, which is only babeled (for SSR) once.

To keep SSR when developing, we would need to always restart @babel/node on changes to ‘src/‘. The packabe ‘babel-watch’ does this, but it’s no longer maintained and incompatible with Babel 7.

All in all I think change is for a better Developer Experience.